[Refactor] 내부 로직을 별개의 함수로 모듈화함
클라우드에 존재하는 기존의 파일과 이름이 곂치지 않는 경로를 만드는 로직을 함수로 분리했다.
